api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic[email]
python-jose
passlib[bcrypt]
faker
langchain
langchain_core
langchain-openai
numpy
httpx
aiohttp
langgraph
apscheduler
colorlog